DARWIN: Use -Werror=objc-method-access for objc
The default Xcode compiler options are insufficient to fail the build if
a method is not found. For example, a typo in the name of method
'componentsJoinedByString' results in the following build warning:

/TaskUtil.m:70:62: warning: instance method '-componentsJoinedBySting:'
          not found (return type defaults to 'id') [-Wobjc-method-access]
          NSLog(@"Task failed: %@ %@ status:%d.", taskName,
                [args componentsJoinedBySting:@", "], status);

Because this is only flagged as a warning, the build completes
successfully.  When this code runs, the AFSBackgrounder merely logs a
runtime exception every time it passes through the erroneous code.  This
is a silent failure, unless you happen to know how to check for
AFSBackgrounder log messages:

  $ log show --predicate 'process=="AFSBackgrounder"'

Add compiler flag '-Werror=objc-method-access' to all of our xcode-built
project files to treat this case as an error instead of a warning, so
the build will fail.

[adeason@sinenomine.net: Update all xcode project files.]

Change-Id: I69a9f45deb6710a50590bd79daf07466332a6ad1
Reviewed-on: https://gerrit.openafs.org/14586
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>