rx: Define symbolic names for Rx magic numbers
Since the original IBM code import, the magic number '6' has been used
to specify the number of consecutive lost keepalives (ping acks) that
indicate a dead connection.  By implication, this also defines the
minimum number of seconds (at the minimum keeplive periodicity of 1
second) before a connection may be considered dead.

Define and use symbolic names for both uses of '6', and document their
relationship.  Both have the same assigned value '6', but
RX_PINGS_LOST_BEFORE_DEAD is use as an ordinal count, while
RX_MINDEADTIME is expressed in units of seconds.

The magic number '12' is used in a couple of places for the default
value of rx_connDeadTime.  Give this constant a name
(RX_DEFAULT_DEAD_TIME) and use it.

No functional change is incurred by this commit.

Change-Id: I8689062d25b51c557da16e27601297b6856e048a
Reviewed-on: https://gerrit.openafs.org/14621
Reviewed-by: Benjamin Kaduk <kaduk@mit.edu>
Tested-by: Benjamin Kaduk <kaduk@mit.edu>