auth: Add 'auser' arg to GenericAuth()
In preparation for a future commit, add an argument (auser) to
GenericAuth, which is the identity of the user associated with the
ticket to be created by this function. Also, introduce a new 'goto done'
destructor to make sure that all code paths go through the same cleanup
code.

Since the new argument is not used by any callers yet, no functional
change is incurred by this commit.

Change-Id: I2f01abc00998baa4f65c2a30143aea93bc1a731d
Reviewed-on: https://gerrit.openafs.org/14596
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>