ptserver: Return error when exceeding _MAXPRLIST
Currently, if the ptserver tries to reply to a client with a 'prlist'
longer than OPENAFS_MAXPRLIST, the call will abort with
RXGEN_SS_MARSHAL (for e.g., PR_GetCPS, PR_ListElements). This is a
rather confusing error code for the peer, since the peer hasn't done
anything wrong, and the error code looks like a networking error since
it's negative.

We can easily avoid this if we just check to see if we are exceeding
the OPENAFS_MAXPRLIST limit, and return a clearer error code. So,
return PRTOOMANY when we extend a 'prlist' to longer than
OPENAFS_MAXPRLIST.

Actually hitting the OPENAFS_MAXPRLIST limit is generally unlikely,
but can be very confusing when it happens. Also log a message in this
situation, to try and help explain what's going on.

Change-Id: Iec03c885989092efad92442bcfe71e1e2ac9f2ac
Reviewed-on: https://gerrit.openafs.org/14732
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>