ptserver: Avoid 'pts adduser' on excessive entries
If we have a user in at least OPENAFS_MAXPRLIST groups, trying to use
that entry will result in PR_GetCPS returning an error, since we
cannot have a 'prlist' with more than OPENAFS_MAXPRLIST entries. But
such an error is not discovered until a PR_GetCPS or similar RPC
actually fails, which means the user is effectively unusable.

So, to catch this earlier, return a PRTOOMANY error when someone tries
to 'pts adduser' for such a user, refusing to create such an
excessively-huge entry in the prdb. This isn't perfect, since a
PR_GetCPS call may still fail if supergroup CPS expansion or a host
CPS puts us over the limit, but this can help catch the most common
cases.

Note that we do not prevent creating a group with over
OPENAFS_MAXPRLIST users in it, even though doing so would make some
RPCs like PR_ListElements unusable for that group. In that situation,
such a group is still generally usable (since GetCPS can still work),
so we don't need to prevent them.

Change-Id: I8857699b34f7f89c64c561f0d3245dde2de893d2
Reviewed-on: https://gerrit.openafs.org/14733
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>