Avoid rxi_tracename overflow
When processing the -trace option for the vlserver (and a couple of
other places), we can easily overflow the rxi_tracename array if the
given string is too big. While the way this global setting works in
general isn't the best, at least for now just prevent the buffer
overflow by doing a simple bounds check with strlcpy.

Change-Id: I41faec8d2aa09f871a69d7db1643f1117aa5618c
Reviewed-on: https://gerrit.openafs.org/14753
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>