vos: Fix vos dump and restore dump file close error messages
The file handle close error messages in the vos dump and restore
commands have "STDIN" and "STDOUT" swapped.  Just remove the filename or
"STDIN"/"STDOUT" from those messages.

Remove the extra flag to indicate if we have a file handle and just
initialize the usd file handles to NULL.

This is a clean up in preparation for other changes.

Change-Id: I97bf5a8177269331e4060746736a882892b9062e
Reviewed-on: https://gerrit.openafs.org/14756
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>