afs: Log failures when mounting /afs
Currently, there are several scenarios where the initial mount of /afs
fails, but we don't log any indication of what the error was. For
non-dynroot, this can happen when the primary cell or root volume
don't exist, or more generally when trying to lookup the root vcache
fails with any error that we don't normally log (most non-network
errors).

Try to log at least something for as many of these scenarios as
possible. This commit covers the more plausible errors where the given
cell/volume don't exist, but also tries to cover some cases where
various internal functions fail (like Linux's bdi_init(), or
afs_CheckInit()).

Some scenarios may not result in the clearest error messages, but at
least some information is logged that can point the user in the right
direction. Avoid logging lookup errors after initial startup
(afs_initState is 200 or greater), to avoid flooding the kernel log
for simple errors like a mispelled volume in a mountpoint.

For example, if the primary cell doesn't exist, we now log something
like this:

    kernel: Starting AFS cache scan...found 0 non-empty cache files (0%).
    kernel: afs: Error getting root volume: cannot get primary cell
    kernel: afs: afs_CheckInit failed with code 110 (afs_initState 200)
    kernel: afs: afs_root failed with code 110

If the root volume doesn't exist, we log something like this:

    kernel: Starting AFS cache scan...found 0 non-empty cache files (0%).
    kernel: afs: Failed to lookup volume root.notexist (code 363524)
    kernel: afs: Failed to lookup volume root.notexist.readonly (code 363524)
    kernel: afs: Error getting root volume: cannot find volume root.notexist
    kernel: afs: afs_CheckInit failed with code 110 (afs_initState 200)
    kernel: afs: afs_root failed with code 110

Change-Id: I31c7a33a40313fd3365fd95ce29d470a9cfad3b4
Reviewed-on: https://gerrit.openafs.org/14885
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>